+++ .env
@@ -2,7 +2,7 @@
 
 # Demo App version
 IMAGE_VERSION=2.0.2
-IMAGE_NAME=ghcr.io/open-telemetry/demo
+IMAGE_NAME=hrexed/open-telemetry/demo
 DEMO_VERSION=latest
 
 # Build Args
@@ -23,7 +23,7 @@ VALKEY_IMAGE=valkey/valkey:8.1-alpine
 TRACETEST_IMAGE=kubeshop/tracetest:${TRACETEST_IMAGE_VERSION}
 
 # Demo Platform
-ENV_PLATFORM=local
+ENV_PLATFORM=linux/amd64
 
 # OpenTelemetry Collector
 HOST_FILESYSTEM=/
