+++ components/pubsub.yaml
@@ -1,0 +1,22 @@
+apiVersion: dapr.io/v1alpha1
+kind: Component
+metadata:
+  name: pubsub
+spec:
+  type: pubsub.kafka
+  version: v1
+  metadata:
+  - name: brokers # Required. Kafka broker connection setting
+    value: "kafka:9092"
+  - name: authType # Required.
+    value: "none"
+  - name: disableTls # Optional. Disable TLS. This is not safe for production!! You should read the `Mutual TLS` section for how to use TLS.
+    value: "false"
+  - name: publishingScopes
+    value: "checkout=orders"
+  - name: subscriptionScopes
+    value: "accounting=orders;fraud-detection=orders"
+scopes:
+- accounting
+- checkout
+- fraud-detection
