+++ src/cart/src/Program.cs
@@ -21,11 +21,7 @@ using OpenFeature.Contrib.Hooks.Otel;
 
 var builder = WebApplication.CreateBuilder(args);
 string valkeyAddress = builder.Configuration["VALKEY_ADDR"];
-if (string.IsNullOrEmpty(valkeyAddress))
-{
-    Console.WriteLine("VALKEY_ADDR environment variable is required.");
-    Environment.Exit(1);
-}
+
 
 builder.Logging
     .AddOpenTelemetry(options => options.AddOtlpExporter())
@@ -33,7 +29,8 @@ builder.Logging
 
 builder.Services.AddSingleton<ICartStore>(x=>
 {
-    var store = new ValkeyCartStore(x.GetRequiredService<ILogger<ValkeyCartStore>>(), valkeyAddress);
+    var store = new DaprStateManagementCartStore(x.GetRequiredService<ILogger<DaprStateManagementCartStore>>(), "cart-state-store");
+
     store.Initialize();
     return store;
 });
@@ -48,7 +45,7 @@ builder.Services.AddSingleton<IFeatureClient>(x => {
 builder.Services.AddSingleton(x =>
     new CartService(
         x.GetRequiredService<ICartStore>(),
-        new ValkeyCartStore(x.GetRequiredService<ILogger<ValkeyCartStore>>(), "badhost:1234"),
+        new DaprStateManagementCartStore(x.GetRequiredService<ILogger<DaprStateManagementCartStore>>(), "bad-state-store"),
         x.GetRequiredService<IFeatureClient>()
 ));
